(finished) product
Picture of the working touchlight. Lego housing contains the arduino and most of the wiring. Quarter, nickel and penny are attached to legos with hot glue. I probably should have soldered the coins on the bottom so that the wires aren't visible but that's what v2.0 is for. With three coins I can have up to seven actions (individual and combination touches).